Remove tty from docker run command
This was breaking the travis job since travis jobs are not tty enabled. Also, using tty here was not working as expected since the container did not always get killed. Instead, this uses the atexit module to execute a docker kill command on the started container(s) when a signal to exit happens.