Support environment variables in task definitions
The most important bit about this PR is that it moves the parsing of the
environment variables to the first operation rather than after it is
validated.

While it certainly will perform better to isolate the job and its tasks
first and then only parse those subsections of the document rather than
the entire document, I found that calling the parsing functions more
than once leads to problems, the most significant of which is that not
every environment variable will be parsed.

For example, in order for an env var to be parsed, it must have its tag
set to the one passed to the yaml module when calling both
add_implicit_resolver and add_constructor. In this case, it's "!path".
The PyYAML's ScalarNode type should have its default tag swapped out for
this one, resulting in a dict such as { "tag": "!path", "value": "$HOME
is where the hate is" }.

However, this doesn't happen when mule's env var parsing function is
called a second time. The tag is set (or not) from the default in
composer.py#compose_scalar_node (the default tag value is
tag:yaml.org,2002:str). Again, if the "!path" key to the resolver's
matrix of regular expressions isn't set correctly here, then the
environment variable in the scalar node that is being examined will
never be parsed (i.e., the constructor passed to the yaml module's
add_constructor function is never called).

I spent quite a bit of time stepping through the PyYAML library until I
finally decided that it's better to parse the entire document up front.