Fix docker ensure
Workdir was being used for the docker build context path. The workDir field is used to say what directory on the created container a mule task should be on, rather than what directory docker should build an image from. I added a context field where this setting can be added in the future.